card footer changes and refactors
