+++ epictrack-web/src/components/shared/userMenu/UserMenu.tsx
@@ -51,7 +51,8 @@ const UserMenu = (props: UserMenuProps) => {
   return (
     <Menu
       sx={{
-        mt: "2.5rem",
+        // mt: "2.5rem",
+        mt: "28px",
         ...props.sx,
       }}
       id={id || "menu-appbar"}
